Fix issues with PHP 4.4
This update fixes issues with PHP 5.4 by removing the typing of
string and integer variables, plus return types.  Unfortunately,
PHP 5.4 has this weird thing were it things a string isn't a string
purely beacuse a scalar string isn't an object string (which answers
that age old question of when is a string not a string, when its a
string)

Also added support for systems where the readline library may not
be enabled within PHP so we can't do a proper countdown in the test
script for the scheduler library.  It now pauses for half a second
before assuming space was pressed.